EnduraCode Main - Merge PREVIOUS pull requests first (#63)
* Created LocalDateTools

For now it just forwards calls to DateTimeTools.

* Updated ReSharper settings

* Created ICalendarTools

And added dependency on iCal.NET library.

* ParsedLine explicit implementation

Modified ParsedLine classes to use explicit member implementation when possible. Also reformatted code.

* Removed obsolete CsvLineParser constructor

There was no way for client code to use it, and its code was exactly duplicated in TextBasedTabularDataParser.buildColumnHeadersToIndexesDictionary.

* Simplified TextBasedTabularDataParser.Parse

It's now parseLine and just returns field data, which reduces duplication in implementations.

* Made TextBasedParsedLine.lineNumber immutable

* Improved TextBasedParsedLine.fields encapsulation

* Made columnHeadersToIndexes immutable

* Cleaned up CsvLineParser

Removed unnecessary members and reformatted.

* Cleaned up FixedWidthParser

Fixed nullability warnings and reformatted.

* Cleaned up TextBasedTabularDataParser

Fixed nullability warnings and reformatted.

* Moved hasHeaderRow setting into TabularDataParser

* Implemented TabularDataParser required columns

* Annotated StringTools.IsNullOrWhiteSpace

Fixes some nullability warnings.

* Implemented required columns in ExcelParser

* Backported ExcelParser improvements

Backported improvements from last commit into TextBasedTabularDataParser and TextBasedParsedLine.

* Fixed tests

* Formatted TabularDataParser

* Renamed ValidationResult to ValidationError

In preparation for a new, public ValidationResult class that will be returned by all the Validator methods.

* Restructured underlying validation methods

Restructured and simplified underlying validation methods in Validator. They now return the error, set the value via a setter parameter, and no longer trim or check for empty.

* Renamed validation method input parameters

* Overhauled Validator.isEmpty to handle trimming

* Re-introduced ValidationMethod delegate

* Overhauled ExecuteValidation to return a result

* Updated public validation methods to return result

* Updated Validator usages elsewhere in TEWL

* Named usages of ExecuteValidation emptyValue param

* Removed result storage from ValidationErrorHandler

* Removed ValidationPackage

This hasn't been used in years, or even decades.

* Removed ValidationError.NoError.

* Removed ErrorCondition.NoError.

* Renamed ErrorCondition to ValidationErrorType

* Formatted ValidationErrorHandler

* Made ExecuteValidation handler parameter nullable

* Made ValidationErrorHandler parameters nullable

* Documented ExecuteValidation handler parameter

* Moved text validations to their own file

* Moved phone validations to their own file

* Moved numeric validations to their own file

* Moved boolean validations to their own file

* Moved ZIP code validations to their own file

* Moved date/time validations to their own file

* Cleaned up Validator

* Converted Error class to a nested record

* Renamed GetNumber to GetNumericString

This clarifies that it is a text-related validation method and should not be lumped in with GetInt, GetShort, etc.

* Disambiguated ValidationError classes

* Fixed doc typo in ValidationResult

* Reversed param order of LineProcessingMethod

THis makes it more consistent with the validation methods of EWL web-framework controls.

* Added UTF-8 BOM in GetTextWriterForWrite

* Added includeBom param to GetTextWriterForWrite

Passing false preserves the behavior the method had before the previous commit.

* Created DurationTools with ToSecondsPhrase method

* Moved in RateLimiter from EWL

This code is tested and already in use by EWL.

* Created PatternString and deprecated IsLike

PatternString enables specialized search controls in EWL and provides a good place to support wildcard characters in the future.

* Moved in TrustedHtmlString from EWL

This has been in use by EWL for years.

* Trivial: Formatted ReflectionTools

* Trivial: Formatted EnumTools

* Fixed nullability warnings in ReflectionTools

* Added type constraints to EnumTools methods

Support for these was added in C# 7.3 in 2018.

* Simplified EnumTools.ToEnglish

* Added IsDefined check in ToEnum

* Expanded LocalDateTools with new methods

Added IsBetween and MonthBeginDate.

* Added LocalDateTools.WeekBeginDate

* Added LocalDateTools.ToMonthYearString

* Added ExecuteWithMachineExclusiveAccess

This is existing code from EWL, but recently improved based on the listed Stack Overflow answer.